Files added to help merge later
Currently, visuals for all the files have not yet been updated to fit react style. We currently have a problem in uploading, where the getStopListUpload function called in queueLoad is returning null instead of the output we need. Thus, we are unable to load the documents required and the site, and are unable to test the other portions of the code.